+++ dhlmex/resources/__init__.py
@@ -1,3 +1,5 @@
-__all__ = ['Resource']
+__all__ = ['Guide', 'PostCode', 'Resource']
 
 from .base import Resource
+from .guides import Guide
+from .post_codes import PostCode
